refactor routers tests
