+++ services/src/test/java/io/scalecube/services/routings/examples/RandomCollection.java
@@ -1,4 +1,4 @@
-package io.scalecube.services.a.b.testing;
+package io.scalecube.services.routings.examples;
 
 import java.util.NavigableMap;
 import java.util.TreeMap;
